+++ packages/create-vlib/template-ts/docs/.vitepress/theme/Layout.vue
@@ -1,0 +1,6 @@
+<!-- .vitepress/theme/Layout.vue -->
+<template>
+  <article style="margin: auto; max-width: 1000px">
+    <Content /><!-- make sure to include markdown outlet -->
+  </article>
+</template>