+++ packages/create-vlib/template-ts/docs/.vitepress/theme/index.ts
@@ -1,0 +1,16 @@
+import DefaultTheme from 'vitepress/theme';
+import DemoBlock from '@ruabick/vitepress-demo-block';
+import '@ruabick/vitepress-demo-block/dist/style.css';
+import './var.css';
+import Layout from './Layout.vue';
+
+export default {
+  ...DefaultTheme,
+  Layout,
+  enhanceApp({ app, router, siteData }) {
+    // app is the Vue 3 app instance from `createApp()`.
+    // router is VitePress' custom router. `siteData` is
+    // a `ref` of current site-level metadata.
+    app.component('demo', DemoBlock);
+  },
+};