+++ packages/create-vlib/template-ts/src/index.ts
@@ -1,0 +1,6 @@
+/**
+ * Example function
+ */
+export function add(a: number, b: number) {
+  return a + b;
+}